Created presenter for Form
